tests: ebpf_perf_drv: minor alteration for xdpdrv
The test originally inspects the `bpf_app1_pkts` and `bpf_app3_pkts`
fields in the output of `ethtool -S`. However these fields are only
relevant when the xdp is loaded in offload mode and represents packets
that are dropped or aborted. When in drv mode there is no distinction
between packets that are dropped and packets that are aborted.
Detecting dropped/aborted packets in drv mode can simply be done by
inspecting the value of `dev_tx_pkts`, if zero then the xdp failed and
dropped or aborted the packets.
The code is altered to only inspect these values if in offload mode.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>