tests: doc: more detailed documentation for test env
Previously the README document for the nfp-drv-kmods-private/tests
provided a basic set of instructions for setting up a test environment.
The README, however, was still limited and lacked detailed
information regarding the specific firmware needed for certain test groups.
The changes made in this commit provide a more extensive explanation
on how to properly set up a testing environment for the driver tests.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Daniel de Villiers <daniel.devilliers@corigine.com>
Reviewed-by: Fei Qin <fei.qin@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>