workflow: update Ubuntu 20.04 hwe path
Turns out that the Ubuntu 20.04 hwe job was looking at the wrong path
for relevant hwe kernels. Expand the current conditional check to
distinguish between Ubuntu 18.04 and 20.04, which allows for different
search paths. For the 20.04 headers the naming style also changes
slightly, so include this in the 'wget -A' flags.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Signed-off-by: Yinjun Zhang <yinjun.zhang@corigine.com>