workflow: allow warnings to fail workflow
In the past, some warnings have not triggered a failure even though the
`-Werror` flag has been set. This addition allows the workflow to detect
warnings as errors. The output of the build will be monitored for
a "warning" keyword, which is not case sensitive. If a "warning" is
found, the script will output a "return 2" and cause a error in the
workflow. This will allow any warnings in the output to trigger errors.

Signed-off-by: Danie du Toit <danie.dutoit@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>