nfp: VF rate limit fix of backport bug
The original backport of
Commit: e0d0e1fdf1ed ("nfp: VF rate limit support")
was done incorrectly. The hook was accidentally applied
to nfp_nfdk_netdev_ops, instead of nfp_nfd3_netdev_ops
like it is in the upstream patch. Move this so the
out-of-tree path matches the in-tree path once again.

Signed-off-by: Bin Chen <bin.chen@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>