compat: updates for kernel 4.18.0-348 on CentOS 8.5.2111
Fix the compile error in nfp_net_compat.h which is
"TCA_PEDIT_KEY_EX_CMD_SET undeclared" when compiling
netro.ko(RDMA driver).

TCA_PEDIT_KEY_EX_CMD_SET is defined in tc_pedit.h
which is included in the condition as below:
 #if (defined(COMPAT__HAVE_METADATA_IP_TUNNEL) ||  \
     LINUX_VERSION_CODE >= KERNEL_VERSION(5, 1, 0))

but the TCA_PEDIT_KEY_EX_CMD_SET is used in function
which is in #else branch as below:
 #if VER_NON_RHEL_LT(5, 1) || VER_RHEL_LT(8, 1)
 ...
 #else
 ...
 #endif

When compiling netro.ko on CentOS 8.5.2111(kernel
4.18.0-348), tc_pedit.h is not included, because
the linux version(4.18.0) is less than 5.1.0. But
the function which uses the TCA_PEDIT_KEY_EX_CMD_SET
will be compiled, because the RHEL version(8.5.2111)
is greater than 8.1.0

Why compiling nfp can be successful, because in the
makefile of nfp COMPAT__HAVE_METADATA_IP_TUNNEL is
defined.

Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Reviewed-by: Yinjun Zhang <yinjun.zhang@corigine.com>
[Louis: commit message on public branch will exclude mention of RDMA]
Signed-off-by: Louis Peens <louis.peens@corigine.com>