packaging: rpm: update DKMS version in spec file
The DKMS version in the spec file currently uses only the major version.
This causes that upgrading completely removes the nfp.ko module, because
the post install script of the old minor version removes the new nfp.ko
module. This is because the path for the modules are the same (only
includes the major version in the path).

The version field is updated to the format version-release, to
differentiate between minor releases.

Signed-off-by: Walter Heymans <walter.heymans@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>