tests: unit: ifstats_reconfig: wait for link stability
The main test loop does not wait for DUT link stability
before performing a TCP ping packet spray after the netdev
channel settings (queue numbers) are changed. This results
in data transmission failure between the EP and DUT, causing
test failure. A link state check was added to the end of
the main loop to wait for DUT link stability.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Reviewed-by: Dano Teixeira <dano.teixeira@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>