workflow: deprecation of ubuntu-18.04 environment in GitHub Actions
The build.yaml workflow made use of the ubuntu-18.04 environment
to run builds against Ubuntu 18.04 kernels before, but GitHub
have announced that the ubuntu-18.04 environment was deprecated as
of 2022-08-08.

In order to prevent breakage of our workflow, deprecate/remove
support for Ubuntu 18.04 from the driver. Using CentOS 7.5.1804
containers to replace it for the corresponding kernel version
(from 4.11 to 4.19) and rockylinux 8.5 (from 4.20 to 5.2) in
"build upstream" and deleting 18.04 test in "build ubuntu".

Signed-off-by: Yu Xiao <yu.xiao@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>