tests: common_test: create conversion helper functions
The older BSP versions were in this format:
010217.010217.010325

A check function was used to skip those versions. Now the
new issue is that the newer BSP versions have the following
format:
22.07-0

Thus, this patch creates helper functions to convert this
new BSP format to a new number for comparing BSP versions
as well as converting it back to a BSP version that can be
interpreted correctly by a user.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>