tests: common_test: compare BSP versions correctly
This patch makes use of the helper functions to compare the
BSP versions correctly such that the tests can make use of
it and then decide whether this test should be skipped or not
based on the BSP versions.

This patch also updates the functions to clarify whether the
check is for the BSP versions or for the NSP API versions.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>