tests: unit: fec_modes: set FEC modes using ethtool
Currently during provisioning, BSP tools are not installed on
the end point. This test never failed as this test only runs
on Carbon/FEC supported cards. Thus, instead of provisioning
the endpoint with BSP tools the test is updated to make use of
ethtool to set the FEC modes for the DUT and EP.

Code was also removed as it is no longer needed since the use
of ethtool is used throughout this test now.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>