tests: reboot: check for correct version of flash-boot
As the BSP format has changed the flash-boot-*.bin file
formats have changed. This causes the regex to only find the
'raw' variance of the flash-boot files and excludes the BSP
variance of the flash-boot files. This causes the nsp_version_re
to be out of range as it only finds 'raw' variance.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Co-developed-by: Huanhuan Wang <huanhuan.wang@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>