tests: reboot: flash_arm: update to base32
The reason this patch updates base64 to base32 is that base64
converts "/" characters to "_" characters. This inputs an
empty specified directory file that should be encoded to the
flash-garbage.bin file, resulting in an empty file.

Initially, the specified file directory comes in as
"/opt/netronome/flash/flash-boot.bin" and is therefore changed to
"_opt_netronome_flash_flash-boot.bin". Thus, nothing can be flashed
and the expected failure cannot be found, which fails the test case
early. For more information please see:
https://www.garykessler.net/library/base64.html

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Co-developed-by: Huanhuan Wang <huanhuan.wang@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>