tests: linux_system: wait_online: restart xmlrpc server
During the flash_arm reboot tests, a reboot occurs after a
flash has occurred. This terminates the xmlrpc server on the
DUT and causes the test to fail due to a timeout error. This
patch fixes that by falling back to ssh instead of using an
xmlrpc connection to communicate with the DUT as this breaks
the test due the xmlrpc server being terminated during reboot.
Once, the host turns on, it re-initiates the xmlrpc server on
the DUT.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>