tests: flower: match_vlan_cfi: unset CFI/DEI bit
The use of the VLAN CFI (Canonical Format Indicator) has
changed to DEI (Drop Eligibility Indicator). (NFP and Intel NICs)
or (their drivers) or (firmware) or (kernel) strips the CFI/DEI
bit upon RX. This caused test failure as the CFI bit was expected
to be set. The bit was unset to align with the repurposing from
CFI to DEI.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>