tests: netdev: update bsp_version test
Updating the bsp version check test as BSP version was in hex format
previously 010217.010217.010325, but is now in string format 22.07-0

Skipping if non-release BSP version is used.

Checking if version is correct length and contains numerical values
where necessary.

Commented to explain reason for test and to explain certain lines
of code.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Dano Teixeira <dano.teixeira@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>