tests: move ethtool functionality to LinuxSystem
This patch moves all the ethtool functionality from DrvSystem
to LinuxSystem. This is a cleanup to consolidate the ethtool
helpers in a single class. This also makes sense as ethtool
is a generic system utility, and not specific to the nfp driver,
it can be useful to have these more widely available.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>