tests: unit: fec_modes: create temporary ethtool version fix
There is an issue when setting the ethtool FEC mode to auto
for ethtool versions greater than 5.12. The issue displays the
FEC active encoding as none when the FEC mode is set to auto.
This cause the test to fail early even though the auto FEC mode
is being set like the others. To get past this a temporary fix
is implemented to pass this test for the FEC auto mode case.

When this issue is resolved then this patch can be taken out in
future as it is no longer needed.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>