tests: drv_system: update part number check
Currently this check only returns the part number in this
format from the nfp-hwinfo command for assembly.partno:

AMDA2000-x001(512MB)

Which we do not really care about whereas an update to
nfp-hwinfo introduces nffw.partno from the nfp-hwinfo
command in this format:

AMDA2000-0001

Which provides a much nicer output for checking. Thus, this
patch makes use of this check if it exists otherwise fallback
to the previous check to obtain the part number.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>