tests: unit: update search for 25G cards
Previously, Carbons were our only 25G cards at the time.
Since then, more 25G cards have been made and thus, a more
robust way of checking for 25G cards is implemented by
discarding the '-XXXX` part of the AMDA number and only
checking for the initial part as there are many variations
for certain cards with the use of a helper function.

Signed-off-by: Dano Teixeira <dano.teixeira@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>