tests: prepend firmware type to tests
The tests.unit and tests.netdev test groups both contain a number of
tests which require different firmware. In order to run only the tests
that make use of a specific firmware, that firmware type name is
prepended to the test name so that all tests using that firmware within
the group can be run using a regex.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>