tests: include dma flag for nfd3+kestrel
When making use of Kestrel firmware which is NFD3-based,
a DMA flag must be specified when loading the driver kernel module.
The absence of this flag leads to the inability to find the
CPP handle when testing on a Kestrel NFP. The insmod function
checks for Kestrel NFPs before adding the flag. An NTI config
file entry was added to control this logic for driver loading.
The test readme was updated to reflect this addition.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Suvana Rohanlal <suvana.rohanlal@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>