tests: netdev: update bsp_version test revision
If the revision number is omitted in the BSP version, it wants to check if
a variable is a number without the variable being created. The revision
number is defined early and only if it is overwritten, it would check
whether it is a digit.

Unique error added if revision number is present and digits are not
numerical.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>