tests: flower: flower_max_entries: determine max from rtsym
The flower_max_entries test made use of a hard-coded max
number of entries set to 480k. This is beyond the
capabilities of Kestrel cards, leading to test failure.
The test now determines the maximum number of entries from
the `CONFIG_FC_HOST_CTX_COUNT` rtsym. A factor is included
to allow scaling based on the returned value.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>