workflows: ci-test-dev: cleanup unused fw_type
Removal of the NIC:agilio-sriov-firmware fw_type as it is not used. And a
change to the inclusion of CoreNIC_tests to reflect this.

Moving the transformation of the dashes in the name of the kernel to
remove to underscores to the its use in the step to uninstall the kernel,
as when transformed before this it would then always be different to the
kernel to reset to, even if they are the same kernel, and so the step
to uninstall would always be triggered.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>