tests: netdev: netconsole: fix pid kill bug
The netconsole test fails since signal -HUP is used and the
script waits for the process to be killed. -HUP is a signal
hangup and does not reliably terminate the process as requested
by the test.

Therefore, the -HUP signal is replaced with -SIGTERM which is
used to terminate a process.

Signed-off-by: Suvana Rohanlal <suvana.rohanlal@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>