packaging: implement prerelease branch functionality
Functionality is implemented to allow creating prerelease branches that
can be used to test possible release candidates. The builds and
automated tests are triggered by creating a branch with the following
format:
- prerelease-YY.MM.rev-rcX

The binaries are uploaded to permanent Azure storage in a 'prereleases'
directory. This will allow testing builds without overwriting previous
release builds that had problems.

The upload location of binaries are also updated to not create a
redundant directory with the same name as the packages.

Signed-off-by: Walter Heymans <walter.heymans@corigine.com>
Reviewed-by: Charles Murcott <charles.murcott@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>