tests: samples: prebuild sample fw for nfp-38xxc/4xxx-b0
Some unit tests make use of sample firmware purpose built
for testing. This firmware is normally built for a specific
chip by specifying the CHIP env variable. This necessitates
rebuilding the firmware when switching between Kestrel and
Osprey cards. The main and sample firmware Makefiles have
been modified to build sample firmware for both chips, and
the tests adapted to make use of the correct fw based on the
card being tested.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Acked-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>