tests: netdev: add ethtool -p test
Implementation of a unit test to ensure the functionality of the
identify function of ethtool by checking the value of the cpld register
using the nfp-cpld BSP tool.

The test will pass if the value of the register is correct when running
the `ethtool -p` command.

The test will fail if there is no change to the cpld register or if the
value of the cpld register is incorrect.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Suvana Rohanlal <suvana.rohanlal@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>