tests: netdev: test_ethtool: expand cleanup code
The test initially used CommonNetdevTest, however this caused
failures when run with other netdev tests. It has now been
changed to CommonTest.

The link test sets the port down but does not set it back up.
A clean-up is added to set the link up for tests that may
follow.

Signed-off-by: Suvana Rohanlal <suvana.rohanlal@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>