tests: common_test: update spawn_vf_netdev function
The spawn_vf_netdev function was cleaned and updated
to return both the vf names and the vf numbers.

The old function required multiple calls depending on the
number of vfs to create. The update allows you to pass
the number of vfs as an input parameter.

The following uses in tests were also updated:
- coalesce_vf
- drv_info
- stats_ethtool

Signed-off-by: Suvana Rohanlal <suvana.rohanlal@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>