tests: sample: bpf: update makefile for llvm
LLVM version 13 can be built on Ubuntu, but the llc commands may
need a <version> to execute.

The Makefile is updated by checking if llc-<version> is in the path,
and if so, to use that path. Else it will use the more standard llc
without the <version> so that LLVM can be compiled.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>