tests: netdev: update AOTC references in phys_port_name test
The flower tests seems to have been previously updated to check for
both the old (AOTC) and new (tc-) flower firmware naming schemes. This
is doing a similar update for the phys_port_name test in the netdev
group.

It also does the check in a slightly more pythonic way, and fixes
up an inconsistent indent.

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>