workflows: ci-test-dev: add quotes around firmware release compare
Having single quotes around the bash if statement causes the variable
to not be expanded. Update this to be double quotes instead, so that
the comparison can work correctly.

Also fix a bug where the "DISA" path was comparing "$RELEASE" against
"release", but just a few lines earlier this actually gets overridden
to "releases".

Signed-off-by: Louis Peens <louis.peens@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>