tests: reload: update path to 10G fw file
Updating the path to the 2x10G firmware file. Previously this was only
available for AMDA0099 cards. This is for the 10G firmware to be used
when the speed has been set to 10G. Using helper function called
get_part_no.

This assumes that the AMDA cards that have the possibility of changing
speeds, have 2 ports.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>