tests: common_test: update bsp_string_to_num conversion
Update bsp_string_to_num to correctly compare multiple different
version formats. This affects the behaviour of check_bsp_min.
This allows tests to be executed on WIP and RC BSP versions, in
addition to the regular release versions.

Additionally, bsp_num_to_string has been removed, because it is
not possible to reconstruct the version string in the case of
a WIP format. The function was also not used anywhere.

Signed-off-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>