tests: common_test: fix kill_pidfile
Previously, kill_pidfile failed silently due to the exit status
of the command being equal to the exit status of the last line,
in this case the exit status of the last if.

This is resolved by invoking a bash shell which can be exited
with the appropriate exit code.

Additionally, the behaviour of the function has been changed to
ensure all specified processes are terminated instead of
allowing some skips.

Signed-off-by: Ryno Swart <ryno.swart@corigine.com>
Reviewed-by: Suvana Rohanlal <suvana.rohanlal@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>