Makefile: fix kmod install signature & System.map
Under Ubuntu depmod fails due to a missing symlink between
the System.map file in the boot directory and the kernel
build directory, as well as a missing signing key. Under
RHEL a warning is shown on each driver module install. This
patch introduces the necessary Makefile changes to solve
these issues.

This improves on a previous attempt which had to be reverted because
of problems on older kernels. This is done by using a certificate
subject of '/CN localhost' which works on a wider range of kernels
compared to the previous value of 'C N=Owner',
which only works on newer kernels.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>