tests: reload: update set_speed test
Update reload.set_speed test to include testing for all cards that
can be set to a different speed. For this, the AMDA number is
retrieved and checked in which speed category the card falls.

The supported speeds are updated.

Part number is no longer used as AMDA numbers are used.

If the port's speed cannot be set to the indicated speed, then the
fail command will now output one of two fail outputs. The check
is updated to see if either of these outputs are seen.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>