tests: use helper function for ip link set up/down
Setting the port links up or down are now using the helper functions
included from the linux_system.py file.

Some commands require fail=False which is now included in the commands
that set the ports up or down.

This ensures that the `ip link set ...` command includes the use of
`dev` as in the official command.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>