workflows: ci-test-dev: incorrect runs-on tag when specifying OS
When selecting a specific OS, the OS name was being passed as a runs-on
tag to the subsequent job, however, no runners exist with just the OS
name as a tag, so the job was never picked up. This fix appends the
necessary prefix to the matrix entry for the tag so that it aligns with
the tags on our GitHub runners.

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>