tests: reload: add reload cleanup
Adding base cleanup applying cleanup of previous classes to reload
tests. Note that base class is implicitly called if the function is not
defined in the derived class. Thus, for these cases, cleanup was called
without base class cleanup. Now cleanup of previous classes is also
called.

This ensures that everything is clear and ready for the next reload
test to run.

Signed-off-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>