tests: drv_system: update get_fw_name() to use load_drv flag
The get_fw_name() function loads and unloads the driver
when running "nfp-media". This is not desirable if the
driver is already loaded within the test before calling
the function.

This patch updates the function to include a "load_drv"
flag that allows the user to specify if the driver needs
to be loaded. This flag is defaulted to "True" as to not
affect current tests that use it.

Signed-off-by: Suvana Rohanlal <suvana.rohanlal@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>