tests: drv_system: insmod: add exception for nfp dependency
Add an exception to the loading of kernel module dependencies
that excludes the NFP kernel driver from being loaded as a
dependency module. The dependency handler for the insmod function
allowed the nfp kernel module to be loaded without proper
parameters when a module is loaded that depends on the NFP
kernel driver.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Reviewed-by: James Hershaw <james.hershaw@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>