tests: common: test_with_traffic: add exp_pkt validation
Add validation of expected packet before attempting to
iterate through a potential NoneType object. Previously in
the case where the function does not fail based on the number
of received packets, but no packets were expected, the helper
attempted to iterate through the NoneType expected packet in
order to log the bad (non-matching) packets, leading to test
failure.

Signed-off-by: Jaco Coetzee <jaco.coetzee@corigine.com>
Reviewed-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Reviewed-by: Ryno Swart <ryno.swart@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>